ref(contract-deployer): <- nake name of that generic
